pyproject.toml
[project]
name = "slack-scraper"
version = "0.1.0"
description = "Scrape multiple Slack workspaces"
requires-python = ">=3.12"
dependencies = [
    "slack-sdk==3.36.0",
    "python-dotenv",
    "requests",
    "google-generativeai>=0.8.5",
    "google-genai>=1.46.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
